Prepare for some code sharing between table and card
