+++ app/javascript/src/collected_inks/components/ArchiveButton.jsx
