+++ app/javascript/src/collected_inks/components/EditButton.jsx
