+++ app/javascript/src/collected_inks/components/InkWithLink.jsx
