+++ app/javascript/src/collected_inks/components/index.js
@@ -1,0 +1,3 @@
+export * from "./ArchiveButton";
+export * from "./EditButton";
+export * from "./InkWithLink";
