+++ app/javascript/src/collected_inks/table/Actions.spec.jsx
@@ -2,7 +2,7 @@
 import React from "react";
 import { render } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
-import { Actions } from "../Actions";
+import { Actions } from "./Actions";
 
 const setup = (jsx, options) => {
   return {
